refactor: rename basic_pid module to core_pid
